Completed webget cpp implementation
Essentially converted the command line inputs from the terminal excercise
to a TCP request, afterwhich I repeatedly  read from the buffer till it
was empty. One challenge was I was printing with cout<<var<<endl, which
would work when ran manually due to the large buffer size, but for the unit
test (which seemed to have a smaller buffer size), would raise errors.